Fix LegacyPeriodDate conversion translation issues
