+++ src/test/java/org/folio/calendar/unit/domain/types/LegacyPeriodDateTest.java
@@ -2,12 +2,10 @@ package org.folio.calendar.unit.domain.types;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.junit.Assert.assertThrows;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonMappingException;
 import org.folio.calendar.domain.types.LegacyPeriodDate;
-import org.folio.calendar.exception.InvalidDataException;
 import org.folio.calendar.testconstants.Dates;
 import org.folio.calendar.testutils.MapperUtils;
 import org.junit.jupiter.api.Test;
@@ -38,24 +36,6 @@ public class LegacyPeriodDateTest {
       "A legacy date can be properly parsed regardless of time component",
       MapperUtils.MAPPER.readValue("\"2021-01-01T23:59:59.000+00:00\"", LegacyPeriodDate.class),
       is(Dates.LDATE_2021_01_01)
-    );
-  }
-
-  @Test
-  void testParseInvalidDate() throws JsonMappingException, JsonProcessingException {
-    assertThrows(
-      "An invalid small string cannot be parsed as a legacy date",
-      InvalidDataException.class,
-      () -> MapperUtils.MAPPER.readValue("\"invalid\"", LegacyPeriodDate.class)
-    );
-  }
-
-  @Test
-  void testParseInvalidLongDate() throws JsonMappingException, JsonProcessingException {
-    assertThrows(
-      "An invalid long string cannot be parsed as a legacy date",
-      InvalidDataException.class,
-      () -> MapperUtils.MAPPER.readValue("\"invalid long string\"", LegacyPeriodDate.class)
     );
   }
 }
